refactor: 랜덤 숫자 이름 power로 변경
