--- src/main/java/model/intgenerator/IntGenerator.java
@@ -1,6 +1,0 @@
-package model.intgenerator;
-
-@FunctionalInterface
-public interface IntGenerator {
-    int pickNumber();
-}
