+++ src/main/java/model/powergenerator/PowerGenerator.java
@@ -1,0 +1,6 @@
+package model.powergenerator;
+
+@FunctionalInterface
+public interface PowerGenerator {
+    int generatePower();
+}
