+++ src/main/java/model/powergenerator/RandomPowerGenerator.java
@@ -1,12 +1,12 @@
-package model.intgenerator;
+package model.powergenerator;
 
 import java.util.Random;
 
-public class RandomIntGenerator implements IntGenerator {
+public class RandomPowerGenerator implements PowerGenerator {
     private static final int NUMBER_BOUND = 10;
 
     @Override
-    public int pickNumber() {
+    public int generatePower() {
         Random random = new Random();
         return random.nextInt(NUMBER_BOUND);
     }
